fix: Resolve signed integer overflow in timestamp test
Changed volatile int to int64_t in TimestampLatencyMeasurement test
to prevent overflow when calculating sum of squares up to 9999^2.

This fixes the UndefinedBehaviorSanitizer error:
runtime error: signed integer overflow: 2146682110 + 3463321
cannot be represented in type 'int'

+++ tests/unit/test_timestamp.cpp
@@ -289,7 +289,7 @@ TEST_F(TimestampTest, TimestampLatencyMeasurement) {
         Timestamp start = current_timestamp();
 
         // Simulate some work - increase workload to ensure measurable time
-        volatile int sum = 0;
+        volatile int64_t sum = 0;
         for (int j = 0; j < 10000; ++j) {
             sum += j * j;  // More complex operation
         }
